modules/swcore: fix rcv_fsm conditional to prevent false force free
Without this fix, when previous frame was whole dropped, transfer fsm was still
in DROP state, while RCV fsm was going to S_IDLE. In that case, RCV fsm saw
tp_drop high and it was doing force_free of a correctly received frame (the one
before dropped frame).
The conditional statement is needed, but it had to be fixed to work according to
description in the comment - i.e. dropping frame when RTU decision is very late.