modules/swcore: prevent transfer fsm from hanging in DROP when force free was not requested
In the situation when RCV fsm was dropping the whole frame (no pages were used)
it was not raising force_free request. However, transfer fsm always waits for
force free to be done once it gets to DROP state. This commit adds rp_drop_no_ff
signal which is asserted by RCV fsm when it goes to DROP, but does not request
force free.