modules/swcore: prevent TR FSM from being stuck in DROP when new SOF comes before LL EOF write of a previous frame
It happened that we had only one clock cycle break between EOF and SOF
for a new frame. In that case LL FSM was not detecting SOF and was not
going to SOF_ON_WR (SOF was already _low_ when LL FSM was in S_WRITE).
If the new frame was decided to be droped after writing only the first
data word to MPM, then we were never storing the new start page to the
FSM and a force_free request was masked. Transfer FSM was not getting
ffree_done and everything was stucked in DROP forever.

The fix adds in_pck_sof_d0 signal to let LL FSM go to S_SOF_ON_WR in the
situation described above. It also keeps _high_ state of
new_pck_first_page signal so that the new start page could be written to
LL when LL FSM goes to IDLE from S_SOF_ON_WR.